+++ packages/status-js/src/index.ts
@@ -1,4 +1,4 @@
-export type { Account } from './account'
+export type { Account } from './client/account'
 export type { ChatMessage as Message } from './client/chat'
 export type { Client, ClientOptions } from './client/client'
 export { createClient } from './client/client'
